==> 01_transform_and_cleanse.ipynb <==
Import relevant libraries to read xls and to transform data

import os
import pandas as pd

Start with the cartel_connections file in raw_data

# Read Excel File which contains data on cartels and their corresponding firms
cartel_df = pd.read_excel("./raw_data/cartels/Cartels.xls", sheet_name='Data_on_all_cartelfirms')

# Get column names from the DataFrame
columns = cartel_df.columns
print(columns)

# select only the relevant columns
cartel_df = cartel_df[['Case:', 'Cartel:', 'Entity Name', 'Key Iustin:', 'Key (invented by Iustin)', 'Start:', 'End:', 'Duration', 'Number of Firms', 'Number_Cartel_Firms',
                       'Sector', 'Subsector', 'CartelClassification']]
# Rename columns for consistency
cartel_df = cartel_df.rename(columns={
    'Case:': 'case',
    'Cartel:': 'cartel',
    'Entity Name': 'entity_name',
    'Key Iustin:': 'key_iustin',
    'Key (invented by Iustin)': 'key_invented',
    'Start:': 'start',
    'End:': 'end',
    'Duration': 'duration',
    'Number of Firms': 'number_of_firms',
    'Number_Cartel_Firms': 'number_cartel_firms',
    'Sector': 'sector',
    'Subsector': 'subsector',
    'CartelClassification': 'cartel_classification'
})

cartel_df.head(10)

# get types of columns
print(cartel_df.dtypes)

# change the type of number_of_cartel_firms
cartel_df['number_cartel_firms'] = cartel_df['number_cartel_firms'].astype('Int64')

print('----------------------')

# check the types of columns
print(cartel_df.dtypes)


# save the cleaned DataFrame to a new Excel file
output_file_path = './transformed_data/base/cartels/cartels.xlsx'

if os.path.exists(output_file_path):
    # If the file already exists, remove it
    os.remove(output_file_path)

# Save the cleaned DataFrame to an Excel file
cartel_df.to_excel(output_file_path, index=False)

# Print the path of the saved file
print(f"Cleaned data saved to: {output_file_path}")

Next: The shareholder files (The files are not always in the same format, so some steps had to be done manually (like deleting the first row in Commerzbank.xls))

# define the directory containing the Excel files (for ease of use)
input_file_path = './raw_data/shareholders'
output_file_path = './transformed_data/base/shareholders'

# loop through all the files in the directory
for filename in os.listdir(input_file_path):
    # check if the file is an Excel file
    if filename.endswith('.xls'):
        
        file_path = os.path.join(input_file_path, filename)

        # Read the Excel file
        df = pd.read_excel(file_path, sheet_name=1)
        
        # drop last 3 rows (total, empty)
        df = df[:-3]

        # Drop empty rows
        df = df.dropna(how='all', axis=0)

        # Get relevant columns with % O/S in the name and filing types
        relevant_columns_percent = df.columns[df.columns.str.contains('% O/S')]
        relevant_columns_filing = df.columns[df.columns.str.contains('Filing Type')]

        # Keep only relevant columns
        df = df[['Investor Name', 'Investor Sub-Type'] + list(relevant_columns_percent) + list(relevant_columns_filing)]

        # Change column names to be more readable by renaming % O/S to perc_os_{month_year} and Filing Type to filing_type_{month_year}
        col_list = []

        # for every column in the dataframe
        for col in df.columns:

            # if the column contains % O/S, extract the month and year and rename the column to perc_os_{month_year}
            if '% O/S' in col:
                month_year = col.split(' ')[-1]
                col_list.append(f'perc_os_{month_year}')
            # if the column contains Filing Type, extract the month and year and rename the column to filing_type_{month_year}
            elif 'Filing Type' in col:
                month_year = col.split(' ')[-1]
                col_list.append(f'filing_type_{month_year}')
            # else, rename the column to lowercase and replace spaces with underscores
            else:
                col_list.append(str.lower(str.replace(col, ' ', '_')))

        # rename the columns
        df.columns = col_list

        company_name = filename.split(".")[0]

        # add the company name to the dataframe
        df["company_name"] = company_name

        # save the dataframe to directory
        df.to_excel(os.path.join(output_file_path, company_name + ".xlsx"), index=False)

        # print the company name to see progress
        print(company_name)

print("---Finished---")
