==> 02_cartel_network.ipynb <==
Import relevant libraries for network and data handling / transformation

import networkx as nx
from pyvis.network import Network
import pandas as pd

Read the cartel file and print some metrics to get an overview of the data

# Read Excel File which contains data on cartels and their corresponding firms
cartel_df = pd.read_excel("./transformed_data/base/cartels/cartels.xlsx")

# Count number of unique case numbers, cartels and entities (companies)
number_of_cases = cartel_df["case"].nunique()
number_of_cartels = cartel_df["cartel"].nunique()
number_of_entities = cartel_df["entity_name"].nunique()

print("Number of cases: ", number_of_cases)
print("Number of Cartels: ", number_of_cartels)
print("Number of companies: ", number_of_entities)
print("Number of Records in Table: ", len(cartel_df))

# Print head as an overview over the dataframe
cartel_df.head(20)

Prepare the different nodes and edges for the graph creation

# Create entity list and set
entity_list = []
entity_set = set()  # Used to check if the firm/company was already added

# Create cartel list and set
cartel_list = []
cartel_set = set() # Used for checking if the cartel was already added

# Create an empty list for the edges between cartels and firms
edge_list = []

for _, row in cartel_df.iterrows(): # _ is the index and will not be used
    
    # Entities 
    if row["entity_name"] not in entity_set:

        # add firm to entity list as a dictionary
        entity_list.append({
            "entity_name": row["entity_name"],
            "key_iustin": row["key_iustin"],
            "key_invented": row["key_invented"],
        })

        # Add firm to set for faster lookup
        entity_set.add(row["entity_name"])  

    # Cartels
    if row["cartel"] not in cartel_set:

        # Add cartel to cartel list as a dictionary
        cartel_list.append({
            "cartel": row["cartel"],
            "case": row["case"],
            "start": row["start"],
            "end": row["end"],
            "duration": row["duration"],
            "number_of_firms": row["number_of_firms"],
            "number_of_cartel_firms": row["number_cartel_firms"],
            "sector": row["sector"],
            "subsector": row["subsector"],
            "cartel_classification": row["cartel_classification"]
        })

        # Add cartel to set for faster lookup
        cartel_set.add(row["cartel"])  

    #Edges
    edge_list.append((row["entity_name"], row["cartel"]))


print("Number of Firm-Nodes in List: ", len(entity_list))
print("Number of Cartel-Nodes in List: ", len(cartel_list))
print("Number of Edges: ", len(edge_list))

Create the graph with the previously created lists

# Create a simple graph
G = nx.Graph()

# Add firm nodes with type "company" and bipartite 0 for left sided alignment in visualization
for entity in entity_list:
    G.add_node(entity["entity_name"], type="Company", bipartite=0, key_iustin=entity["key_iustin"], key_iustin_long=entity["key_invented"])

# Add cartel nodes with type "cartel" and bipartite 1 for right sided alignment in visualization
for cartel in cartel_list:
    G.add_node(cartel["cartel"], type="Cartel", bipartite=1, case=cartel["case"], startyear=cartel["start"], endyear=cartel["end"], 
               duration=cartel["duration"], number_of_firms=cartel["number_of_firms"], number_of_cartel_firms=cartel["number_of_cartel_firms"], 
               sector=cartel["sector"], subsector=cartel["subsector"], classification=cartel["cartel_classification"])

# Add edges
G.add_edges_from(edge_list)

Save it for later visualization

# Save the graph as a graphml file for later visualization
nx.write_graphml(G, f"./transformed_data/cartel_networks/cartel_network.graphml", named_key_ids=True, infer_numeric_types=True)

Calculate centrality metrics for cartel and company nodes

# Calculate centrality measures for all nodes
centrality_measures = {
    "degree_centrality": nx.degree_centrality(G),
    "betweenness_centrality": nx.betweenness_centrality(G),
    "closeness_centrality": nx.closeness_centrality(G),
    "eigenvector_centrality": nx.eigenvector_centrality(G),
    "pagerank": nx.pagerank(G),
    "harmonic_centrality": nx.harmonic_centrality(G),
    "degree": dict(G.degree())
}

# Create seperate lists for company/firm and cartel centralities
company_centrality = []
cartel_centrality = []

# Loop through the nodes in the graph
for node, attributes in G.nodes(data=True):

    # Get the node type (Cartel or Company)
    node_type = attributes["type"]

    # Create dictionary with name of company or cartel and their corresponding centrality measures
    centrality_data = {
        "name": node,
        "degree_centrality": centrality_measures["degree_centrality"].get(node, 0),
        "betweenness_centrality": centrality_measures["betweenness_centrality"].get(node, 0),
        "closeness_centrality": centrality_measures["closeness_centrality"].get(node, 0),
        "eigenvector_centrality": centrality_measures["eigenvector_centrality"].get(node, 0),
        "pagerank": centrality_measures["pagerank"].get(node, 0),
        "harmonic_centrality": centrality_measures["harmonic_centrality"].get(node, 0),
        "degree": centrality_measures["degree"].get(node, 0)
    }

    if node_type == "Company":
        company_centrality.append(centrality_data)
    elif node_type == "Cartel":
        cartel_centrality.append(centrality_data)

# Create dataframes for cartel and company centrality
df_cartel_centrality = pd.DataFrame(cartel_centrality).add_prefix('cartel_')
df_company_centrality = pd.DataFrame(company_centrality).add_prefix('company_')

df_cartel_centrality.head()



# Merge the cartel_centrality dataframe
df_merged = pd.merge(cartel_df, df_cartel_centrality, left_on="cartel", right_on="cartel_name", how="left")

# Merge the company_centrality dataframe
df_merged = pd.merge(df_merged, df_company_centrality, left_on="entity_name", right_on="company_name", how="left")

# Drop the company_name and cartel_name column (redundant)
df_merged.drop(columns=["company_name"], inplace=True)
df_merged.drop(columns=["cartel_name"], inplace=True)

# Save merged data to excel
df_merged.to_excel("./transformed_data/cartel_networks/cartel_data_with_centrality_measures.xlsx", index=False)
