==> 05_connected_cartel_network.ipynb <==
Import relevant libraries

import pandas as pd
import numpy as np

Read the cartel file

# read cartel excel file
cartel_df = pd.read_excel('./transformed_data/base/cartels/cartels.xlsx')

# count number of rows in cartel_df
print(f"number of records in cartel_df:  {cartel_df.shape[0]}")

cartel_df.head(10)

Read the mapping file to map entity_name to file_names

# Get all filenames from mapping file (contains all filenames)
df_mapping = pd.read_excel("./raw_data/mapping/filename_to_entity_mapping.xlsx")

# if entity_name_direct_match column is not empty take entity_name_direct_match else take entity_name_indirect_match_1
df_mapping['entity_name'] = df_mapping.apply(
    lambda x: x['entity_name_direct_match'] if pd.notna(x['entity_name_direct_match']) else x['entity_name_indirect_match_1'], axis=1)

# drop all columns except entity_name and filename
df_mapping = df_mapping[['entity_name', 'file_name']]

# join cartel_df with df_mapping on entity_name
cartel_df = cartel_df.merge(df_mapping, on='entity_name', how='left')

# count number of rows in cartel_df
print(f"number of records in cartel_df:  {cartel_df.shape[0]}")

cartel_df.head(10)

Get the amount of file_names per cartel (Count not null per cartel)

# Assuming your DataFrame is named df
# Step 1: Filter non-null file names
cartel_df_not_null = cartel_df[cartel_df['file_name'].notnull()]

# Step 2: Group by 'case' and 'cartel', and count non-null file_name entries
cartel_df_fn_counts = cartel_df_not_null.groupby(['case', 'cartel'])['file_name'].count().reset_index()

# Optional: Rename the column to something meaningful
cartel_df_fn_counts = cartel_df_fn_counts.rename(columns={'file_name': 'number_of_firms_with_file'})

cartel_df_fn_counts.head(10)

Join the counts of the previous step with the cartel file to get the number_of_firms_with_file for each cartel/Case

cartel_df = cartel_df.merge(cartel_df_fn_counts, left_on=['case', 'cartel'], right_on=['case', 'cartel'], how='left')

# fill NaN values with 0
cartel_df['number_of_firms_with_file'] = cartel_df['number_of_firms_with_file'].fillna(0)
# convert to int
cartel_df['number_of_firms_with_file'] = cartel_df['number_of_firms_with_file'].astype(int)

# fill NaN values with 0
cartel_df['number_cartel_firms'] = cartel_df['number_cartel_firms'].fillna(0)
# convert to int
cartel_df['number_cartel_firms'] = cartel_df['number_cartel_firms'].astype(int)

# fill NaN values with 0
cartel_df['number_of_firms'] = cartel_df['number_of_firms'].fillna(0)
# convert to int
cartel_df['number_of_firms'] = cartel_df['number_of_firms'].astype(int)

# change order of columns
cartel_df = cartel_df[['case', 'cartel', 'number_of_firms', 'number_cartel_firms', 'number_of_firms_with_file', 'sector', 'subsector', 'entity_name', 'file_name', 'duration', 'start', 'end', ]]

print(f"number of records in cartel_df_v2:  {cartel_df.shape[0]}")

print(cartel_df.dtypes)

cartel_df.head(10)


Prepare the t_A... columns in the cartel file

# get largest duration value
max_duration = cartel_df['duration'].max()
print(f"max duration:   {max_duration}")

# get records with max duration
print(f"cartel name:    {cartel_df[cartel_df['duration'] == max_duration]["cartel"].unique()}")

cartel_df['t_A_-3'] = cartel_df['start'] - 3
cartel_df['t_A_-2'] = cartel_df['start'] - 2
cartel_df['t_A_-1'] = cartel_df['start'] - 1
cartel_df['t_A'] = cartel_df['start']

# for range 1 to max_duration create new column with name t_A_+1, t_A_+2, t_A_+3 and so on
for i in range(1, max_duration - 1):
# Use np.where to conditionally assign values
    cartel_df[f't_A_+{i}'] = np.where(cartel_df['start'] + i < cartel_df['end'], 
                                         cartel_df['start'] + i, 
                                         None)

cartel_df['t_E'] = cartel_df['end']
cartel_df['t_E_+1'] = cartel_df['end'] + 1
cartel_df['t_E_+2'] = cartel_df['end'] + 2
cartel_df['t_E_+3'] = cartel_df['end'] + 3

cartel_df.head(10)

Prepare the v_A... fields for the corresponding values to each time slot

cartel_df['v_A_-3'] = None
cartel_df['v_A_-2'] = None
cartel_df['v_A_-1'] = None
cartel_df['v_A'] = None

# for range 1 to max_duration create new column with name t_A_+1, t_A_+2, t_A_+3 and so on
for i in range(1, max_duration - 1):
# Use np.where to conditionally assign values
    cartel_df[f'v_A_+{i}'] = None

cartel_df['v_E'] = None
cartel_df['v_E_+1'] = None
cartel_df['v_E_+2'] = None
cartel_df['v_E_+3'] = None

Fill in the v_A... fields with the herfindal index

# crate copy of cartel_df
cartel_df_herfindal = cartel_df.copy()

# get all column names that start with t_A OR t_E
t_a_columns = [col for col in cartel_df_herfindal.columns if col.startswith('t_A') or col.startswith('t_E')]
# get all column names that start with v_A OR v_E
v_e_columns = [col for col in cartel_df_herfindal.columns if col.startswith('v_A') or col.startswith('v_E')]

print(t_a_columns)
print(v_e_columns)

herfindal_df = pd.read_excel('./transformed_data/shareholder_data/company_metrics.xlsx')

def get_herfindahl_value(year, company_name):
    match = herfindal_df[
        (herfindal_df['year'] == year) &
        (herfindal_df['company_name'] == company_name)
    ]['herfindahl_index']
    return match.values[0] if not match.empty else np.nan

for index, row in cartel_df_herfindal.iterrows():

    file_name = row['file_name']

    for col in t_a_columns:
        # get the value of the column
        t_a_value = row[col]
        # check if the value is not null
        if pd.notna(t_a_value):
            # get the herfindahl value for the year and company name
            herfindahl_value = get_herfindahl_value(int(t_a_value), file_name)

            if pd.notna(herfindahl_value):
                # set the value in the corresponding v_A column
                cartel_df_herfindal.at[index, col.replace('t_', 'v_')] = herfindahl_value

print('Finished')

cartel_df_herfindal.head(10)

Save the excel for analysis

# save cartel_df_v2 to excel file
cartel_df_herfindal.to_excel('./transformed_data/cartel_networks/cartel_df_herfindal.xlsx', index=False)

Calculate Ownership of investors on cartels by year

merged_shareholders_df = pd.read_excel('./transformed_data/connected_shareholder_data/merged_shareholder_data.xlsx')

# get all column names that start with perc_os
perc_os_columns = [col for col in merged_shareholders_df.columns if col.startswith('perc_os')]

# select only relevant columns
merged_shareholders_df = merged_shareholders_df[['investor_name', 'company_name'] + perc_os_columns]

merged_shareholders_df.head(10)

# join cartel_df with merged_shareholders_df on file_name = company_name
cartel_df_sh = cartel_df.merge(merged_shareholders_df, left_on='file_name', right_on='company_name', how='left')
# drop company_name column 
cartel_df_sh = cartel_df_sh.drop(columns=['company_name'])

# count rows in cartel_df_sh
print(f"number of records in cartel_df_sh:  {cartel_df_sh.shape[0]}")

t_columns = [col for col in cartel_df_sh.columns if col.startswith('t_A') or col.startswith('t_E')]

for index, row in cartel_df_sh.iterrows():
    
    for col in t_columns:
        # get the value of the column
        t_value = row[col]
        # check if the value is not null
        if pd.notna(t_value):
            try:
                # get the perc_os value for the year and company name
                perc_os_value = row['perc_os_31-Dec-' + str(t_value)]
            except KeyError:
                # if the column does not exist, set perc_os_value to NaN
                perc_os_value = np.nan
            # set the value in the corresponding v_A column
            cartel_df_sh.at[index, col.replace('t_', 'v_')] = perc_os_value    

Make the dataset more lightweight

# drop all perc_os columns
cartel_df_sh = cartel_df_sh.drop(columns=perc_os_columns)
# drop all t_A and t_E columns
cartel_df_sh = cartel_df_sh.drop(columns=t_columns)
# drop all rows with v_A and v_E columns all null
cartel_df_sh = cartel_df_sh.dropna(subset=[col for col in cartel_df_sh.columns if col.startswith('v_')], how='all')

# change order of columns
cartel_df_sh = cartel_df_sh[['case', 'cartel', 'number_of_firms', 'number_cartel_firms', 'number_of_firms_with_file', 'sector', 'subsector', 'duration', 'start', 'end', 'entity_name', 'file_name', 'investor_name'] + [col for col in cartel_df_sh.columns if col.startswith('v_')]]

Save Excel with Ownership of Investors by Cartel and Company

cartel_df_sh.to_excel('./transformed_data/cartel_networks/cartel_df_sh_by_entities.xlsx', index=False)

Prepare Excel for Ownership of Investors only by cartel (Avg Investment per Cartel)

# drop entity_name and file_name columns
cartel_df_sh = cartel_df_sh.drop(columns=['entity_name', 'file_name'])

# get average of all v_A and v_E columns by groupby case and cartel
cartel_df_sh_avg_by_cartel = cartel_df_sh.groupby(['case', 'cartel', 'number_of_firms', 'number_cartel_firms', 'number_of_firms_with_file', 'sector', 'subsector', 'duration', 'start', 'end', 'investor_name']).mean().reset_index()

cartel_df_sh_avg_by_cartel.head(10)

Save Excel with Ownership of Investors by Cartel

cartel_df_sh_avg_by_cartel.to_excel('./transformed_data/cartel_networks/cartel_df_sh_by_cartel.xlsx', index=False)

Prepare Excel for Ownership of Investors over all cartels (averaged)


# drop all columns except v_* columns and investor_name
cartel_df_sh = cartel_df_sh.drop(columns=['case', 'cartel', 'number_of_firms', 'number_cartel_firms', 'number_of_firms_with_file', 'sector', 'subsector', 'duration', 'start', 'end'])

# get average of all v_A and v_E columns by groupby investor_name
cartel_df_sh_avg_by_investor = cartel_df_sh.groupby(['investor_name']).mean().reset_index()

cartel_df_sh_avg_by_investor.head(10)

Save Excel with Ownership of Investors over all cartels

cartel_df_sh_avg_by_investor.to_excel('./transformed_data/cartel_networks/cartel_df_sh_by_investor.xlsx', index=False)
