==> 90_identify_missing_investor_identities.ipynb <==
import pandas as pd

# Read Excel File which contains data on cartels and their corresponding firms
df_entities = pd.read_excel("./transformed_data/base/cartels/cartels.xlsx")

# Get unique entities from the "Entity Name" column
entities = df_entities["entity_name"].unique()

# Get all filenames from mapping file (contains all filenames)
df_mapping = pd.read_excel("./raw_data/mapping/filename_to_entity_mapping.xlsx")

# if entity_name_direct_match column is not empty take entity_name_direct_match else take entity_name_indirect_match_1
df_mapping['entity_name'] = df_mapping.apply(
    lambda x: x['entity_name_direct_match'] if pd.notna(x['entity_name_direct_match']) else x['entity_name_indirect_match_1'], axis=1)

# Get unique entities from the mapping file
mapping_entities = df_mapping['entity_name'].unique()

# count unique entities in the mapping file
mapping_entities_count = len(mapping_entities)

print(mapping_entities)
print(mapping_entities_count)


# define new dataframe with 2 columns entity_name and file_name
df_results = pd.DataFrame(columns=['entity_name', 'file_name'])

# Iterate through each entity to check if it exists in the mapping file
for entity in entities:
    # Check if the entity exists in the mapping file
    if entity not in mapping_entities:
        # add entity and None to the dataframe
        df_results.loc[len(df_results)] = [entity, None]
    else:
        # Get the corresponding file name for the entity
        file_name = df_mapping.loc[df_mapping['entity_name'] == entity, 'file_name'].values[0]
        # add entity and file_name to the dataframe
        df_results.loc[len(df_results)] = [entity, file_name]

# count entities where file_name is not None
df_results_count = df_results[df_results['file_name'].notna()].shape[0]
# count entities where file_name is None
df_results_none_count = df_results[df_results['file_name'].isna()].shape[0]

print(f"Number of Entities with investor file: {df_results_count}")
print(f"Number of Entities without investor file: {df_results_none_count}")
print(f"Number of Entities total: {len(df_results)}")

# save the results to a new Excel file
df_results.to_excel("./temp_data/entities_with_corresponding_filenames.xlsx", index=False)

# for each filename in df_mapping
for file_name in df_mapping['file_name']:
    # check if the file exists in the mapping file
    if file_name not in df_results['file_name'].values:
        # add file_name and None to the dataframe
        print(file_name)


import os

# get all filenames from directory
directory = "./transformed_data/base/shareholders/"
filenames = os.listdir(directory)
# remove .xlsx from filenames
filenames = [filename.replace('.xlsx', '') for filename in filenames]

for filename in filenames:
    # check if the file exists in the mapping file
    if filename not in df_mapping['file_name'].values:
        # add file_name and None to the dataframe
        print(filename)
