==> 91_identify_cartels_with_most_files.ipynb <==
import pandas as pd
import numpy as np

cartel_df = pd.read_excel('./transformed_data/cartel_networks/cartel_df_herfindal.xlsx')

cartel_df = cartel_df[['case', 'cartel', 'number_of_firms', 'number_cartel_firms', 'number_of_firms_with_file', 'start', 'end']]

# select distinct rows
cartel_df = cartel_df.drop_duplicates(subset=['case', 'cartel', 'number_of_firms', 'number_cartel_firms', 'number_of_firms_with_file'])

cartel_df.head(10)

# get rows where number_of_firms_with_file is at least 50% of number_of_firms
cartel_df_filtered = cartel_df[cartel_df['number_of_firms_with_file'] >= 0.8 * cartel_df['number_of_firms']]

# print amount of records before and after filtering
print(f"Number of records before filtering: {len(cartel_df)}")
print(f"Number of records after filtering: {len(cartel_df_filtered)}")

cartel_df_filtered.head(10)

cartel_df_filtered.to_excel('./temp_data/cartels_with_more_than_80_percent_firms.xlsx', index=False)
