==> 92_reduced_company_centrality_metrics.ipynb <==
Calculate the Centrality Metrics for the Cartels where we have at least 80% of the investor files

import pandas as pd
import numpy as np
import os
import networkx as nx
from networkx import PowerIterationFailedConvergence


Identify cartels with at least 80% of companies have a corresponding file

cartel_df = pd.read_excel('./transformed_data/cartel_networks/cartel_df_herfindal.xlsx')

cartel_df = cartel_df[['case', 'cartel', 'number_of_firms', 'number_cartel_firms', 'number_of_firms_with_file', 'start', 'end', 'duration', 'entity_name', 'file_name']]

# get rows where number_of_firms_with_file is at least 50% of number_of_firms
cartel_df_filtered = cartel_df[cartel_df['number_of_firms_with_file'] >= 0.8 * cartel_df['number_of_firms']]

cartel_df_filtered.head(10)

filter the dataframe further to select cartels with more than 1 investor file, because of star-shaped graph problem

# filer out rows where 'number_of_firms' is 1 
cartel_df_filtered = cartel_df_filtered[cartel_df_filtered['number_of_firms'] > 1]


# get distinct file_names
file_names = cartel_df_filtered['file_name'].unique()

df_all = pd.DataFrame()

# get each file in the directory
for file in os.listdir("./transformed_data/base/shareholders"):
    if file.endswith(".xlsx") and file[:-5] in file_names:
        # read the file
        file_path = os.path.join("./transformed_data/base/shareholders", file)
        
        df = pd.read_excel(file_path)

        # union all the dataframes
        df_all = pd.concat([df_all, df], ignore_index=True)

df_all_company_names = df_all['company_name'].unique()
print(df_all_company_names)

cartel_df_filtered_with_sh = cartel_df_filtered.merge(df_all, left_on='file_name', right_on='company_name', how='left', suffixes=('', '_y'), indicator=True)

cartel_df_filtered_with_sh.head(10)

# get all perc_os columns in the dataframe
perc_os_columns = [col for col in cartel_df_filtered_with_sh.columns if 'perc_os' in col]

print(perc_os_columns)

# count unique company names
cartel_names = cartel_df_filtered_with_sh['cartel'].unique()
print(cartel_names)

# order dataframe by cartel
cartel_df_filtered_with_sh = cartel_df_filtered_with_sh.sort_values(by=['cartel'])

# split the dataframe into multiple dataframes based on cartel names
cartel_dfs = []

for cartel_name in cartel_names:
    cartel_dfs.append(cartel_df_filtered_with_sh[cartel_df_filtered_with_sh['cartel'] == cartel_name].copy())


# for each column in perc_os_columns
for col in perc_os_columns:

    # Get year from the column name
    year = col[-4:]

    for single_cartel_df in cartel_dfs:
        
        # get the cartel name
        cartel_name = single_cartel_df['cartel'].iloc[0]

        # Define a new graph for the year e.g. 1997 and cartel name
        G = nx.Graph()

        for index, row in single_cartel_df.iterrows():

            # if value is not null in row[col]
            if pd.notnull(row[col]):
                
                # add node with company name
                G.add_node(row['company_name'], type='Company')
                # add investor name
                G.add_node(row['investor_name'], type='Investor')
                # add edge between cartel and company
                G.add_edge(row['company_name'], row['investor_name'], weight=row[col])

        # save the graph to a file
        nx.write_graphml(G, f"./transformed_data/reduced_connected_shareholder_networks/shareholder_network_{cartel_name}_{year}.graphml", named_key_ids=True, infer_numeric_types=True)

company_measure_list = []

for file in os.listdir("./transformed_data/reduced_connected_shareholder_networks"):
    if file.endswith(".graphml"):
        # create new graph from .graphml file
        G = nx.read_graphml(f"./transformed_data/reduced_connected_shareholder_networks/{file}")
        
        # get the cartel name from the file name
        cartel_name = file.split('_')[2]
        # get the year from the file name
        year = file.split('_')[3].split('.')[0]

        print(f"Calculating centrality measures for {cartel_name} in {year}")

        # calculate the node centralities for each company in the graph
        for node in G.nodes():
            if G.nodes[node]['type'] == 'Company':
                
                # get degree of the node
                degree = G.degree(node)

                # calculate the degree centrality
                degree_centrality = nx.degree_centrality(G)[node]
                # calculate the closeness centrality
                closeness_centrality = nx.closeness_centrality(G, node)
                # calculate the betweenness centrality
                betweenness_centrality = nx.betweenness_centrality(G)[node]

                # calculate the centralities for the company node
                degree_centrality = nx.degree_centrality(G)[node]
                closeness_centrality = nx.closeness_centrality(G)[node]
                betweenness_centrality = nx.betweenness_centrality(G)[node]
                harmonic_centrality = nx.harmonic_centrality(G)[node]
                pagerank = nx.pagerank(G)[node]
            
                try:
                    eigenvector_centrality = nx.eigenvector_centrality(G)[node]    
                    katz_centrality = nx.katz_centrality(G)[node]

                except PowerIterationFailedConvergence:
                    # if the eigenvector centrality calculation fails, set it to None
                    eigenvector_centrality = None
                    katz_centrality = None

                # add measures to dictionary
                measures = {
                    'cartel_name': cartel_name,
                    'company_name': node,
                    'year': year,
                    'degree': degree,
                    'degree_centrality': degree_centrality,
                    'closeness_centrality': closeness_centrality,
                    'betweenness_centrality': betweenness_centrality,
                    'eigenvector_centrality': eigenvector_centrality,
                    'harmonic_centrality': harmonic_centrality,
                    'katz_centrality': katz_centrality,
                    'pagerank': pagerank
                }

                company_measure_list.append(measures)

measures_df = pd.DataFrame(company_measure_list)

# Save to Excel
measures_df.to_excel("./transformed_data/reduced_connected_shareholder_data/reduced_company_node_stats.xlsx", index=False)
